Updated user_proc. Fixed conf problem

+++ src/local_lib/headers/semaphore.h
@@ -8,19 +8,19 @@
  * @param semaphore_index index of the semaphore to exe the lock on
  * @return 0 in case of success, -1 otherwise
  */
-int semaphore_lock(int semaphore_id, int semaphore_index);
+int semaphore_lock(int semaphore_id, short int semaphore_index);
 /**
  * Execute an unlock on sempaphore_index of semaphore_id array
  * @param semaphore_id array of semaphores indexs
  * @param semaphore_index index of the semaphore to exe the lock on
  * @return 0 in case of success, -1 otherwise
  */
-int semaphore_unlock(int semaphore_id, int sempahore_index);
+int semaphore_unlock(int semaphore_id,short int sempahore_index);
 /**
  * Execute a wait for zero on sempaphore_index of semaphore_id array
  * @param semaphore_id array of semaphores indexs
  * @param semaphore_index index of the semaphore to exe the lock on
  * @return 0 in case of success, -1 otherwise
  */
-int semaphore_wait_for_sinc(int semaphore_id, int semaphore_index);
+int semaphore_wait_for_sinc(int semaphore_id, short int semaphore_index);
 #endif /*DGT_SO_PROJECT_21_22_SEMAPHORE_H*/
